--- packages/mask/src/plugins/Trader/apis/trending/mirrored_tokens.json
@@ -1,93 +1,0 @@
-[
-    {
-        "id": "0x21ca39943e91d704678f5d00b6616650f066fd63",
-        "name": "Wrapped Mirror TSLA Token",
-        "symbol": "mTSLA",
-        "decimals": 18,
-        "contract_address": "0x21ca39943e91d704678f5d00b6616650f066fd63"
-    },
-    {
-        "id": "0xd36932143f6ebdedd872d5fb0651f4b72fd15a84",
-        "name": "Wrapped Mirror mAAPL Token",
-        "symbol": "mAAPL",
-        "decimals": 18,
-        "contract_address": "0xd36932143f6ebdedd872d5fb0651f4b72fd15a84"
-    },
-    {
-        "id": "0x41bbedd7286daab5910a1f15d12cbda839852bd7",
-        "name": "Wrapped Mirror mMSFT Token",
-        "symbol": "mMSFT",
-        "decimals": 18,
-        "contract_address": "0x41bbedd7286daab5910a1f15d12cbda839852bd7"
-    },
-    {
-        "id": "0x13b02c8de71680e71f0820c996e4be43c2f57d15",
-        "name": "Wrapped Mirror mQQQ Token",
-        "symbol": "mQQQ",
-        "decimals": 18,
-        "contract_address": "0x13b02c8de71680e71f0820c996e4be43c2f57d15"
-    },
-    {
-        "id": "0x0cae9e4d663793c2a2a0b211c1cf4bbca2b9caa7",
-        "name": "Wrapped Mirror mAMZN Token",
-        "symbol": "mAMZN",
-        "decimals": 18,
-        "contract_address": "0x0cae9e4d663793c2a2a0b211c1cf4bbca2b9caa7"
-    },
-    {
-        "id": "0x1d350417d9787e000cc1b95d70e9536dcd91f373",
-        "name": "Wrapped Mirror mIAU Token",
-        "symbol": "mIAU",
-        "decimals": 18,
-        "contract_address": "0x1d350417d9787e000cc1b95d70e9536dcd91f373"
-    },
-    {
-        "id": "0x56aa298a19c93c6801fdde870fa63ef75cc0af72",
-        "name": "Wrapped Mirror mBABA Token",
-        "symbol": "mBABA",
-        "decimals": 18,
-        "contract_address": "0x56aa298a19c93c6801fdde870fa63ef75cc0af72"
-    },
-    {
-        "id": "0x9d1555d8cB3C846Bb4f7D5B1B1080872c3166676",
-        "name": "Wrapped Mirror mSLV Token",
-        "symbol": "mSLV",
-        "decimals": 18,
-        "contract_address": "0x9d1555d8cB3C846Bb4f7D5B1B1080872c3166676"
-    },
-    {
-        "id": "0x31c63146a635eb7465e5853020b39713ac356991",
-        "name": "Wrapped Mirror mUSO Token",
-        "symbol": "mUSO",
-        "decimals": 18,
-        "contract_address": "0x31c63146a635eb7465e5853020b39713ac356991"
-    },
-    {
-        "id": "0xc8d674114bac90148d11d3c1d33c61835a0f9dcd",
-        "name": "Wrapped Mirror mNFLX Token",
-        "symbol": "mNFLX",
-        "decimals": 18,
-        "contract_address": "0xc8d674114bac90148d11d3c1d33c61835a0f9dcd"
-    },
-    {
-        "id": "0xedb0414627e6f1e3f082de65cd4f9c693d78cca9",
-        "name": "Wrapped Mirror mTWTR Token",
-        "symbol": "mTWTR",
-        "decimals": 18,
-        "contract_address": "0xedb0414627e6f1e3f082de65cd4f9c693d78cca9"
-    },
-    {
-        "id": "0xf72fcd9dcf0190923fadd44811e240ef4533fc86",
-        "name": "Wrapped Mirror mVIXY Token",
-        "symbol": "mVIXY",
-        "decimals": 18,
-        "contract_address": "0xf72fcd9dcf0190923fadd44811e240ef4533fc86"
-    },
-    {
-        "id": "0x59a921db27dd6d4d974745b7ffc5c33932653442",
-        "name": "Wrapped Mirror mGOOGL Token",
-        "symbol": "mGOOGL",
-        "decimals": 18,
-        "contract_address": "0x59a921db27dd6d4d974745b7ffc5c33932653442"
-    }
-]
