+++ packages/web3-providers/src/coingecko/helpers.ts
@@ -1,7 +1,7 @@
 import { createLookupTableResolver } from '@masknet/shared-base'
 import { ChainId } from '@masknet/web3-shared-evm'
 
-export const resolveChain = createLookupTableResolver<string, ChainId | undefined>(
+export const resolveCoinGeckoChainId = createLookupTableResolver<string, ChainId | undefined>(
     {
         ethereum: ChainId.Mainnet,
         'binance-smart-chain': ChainId.BSCT,
