+++ packages/web3-providers/src/coingecko/index.ts
@@ -1,2 +1,3 @@
+export * from './apis/CoinGecko.js'
 export * from './apis/CoinGeckoEVM.js'
 export * from './apis/CoinGeckoSolana.js'
