+++ packages/web3-providers/src/trending/helpers.ts
@@ -1,0 +1,17 @@
+import MIRRORED_TOKENS from './mirrored_tokens.json'
+import type { TrendingAPI } from '../types/index.js'
+
+export function isMirroredKeyword(symbol: string) {
+    return MIRRORED_TOKENS.map((x) => x.symbol).some((x) => x.toUpperCase() === symbol.toUpperCase())
+}
+
+export function getCommunityLink(links: string[]): TrendingAPI.CommunityUrls {
+    return links.map((x) => {
+        if (x.includes('twitter')) return { type: 'twitter', link: x }
+        if (x.includes('t.me')) return { type: 'telegram', link: x }
+        if (x.includes('facebook')) return { type: 'facebook', link: x }
+        if (x.includes('discord')) return { type: 'discord', link: x }
+        if (x.includes('reddit')) return { type: 'reddit', link: x }
+        return { type: 'other', link: x }
+    })
+}
