+++ packages/web3-providers/src/trending/index.ts
@@ -1,0 +1,1 @@
+export {}
