+++ packages/web3-providers/src/trending/mirrored_tokens.json
